+++ src/main/java/com/management/tab/application/auth/LoginService.java
@@ -1,6 +1,6 @@
-package com.management.tab.config.auth.security.service;
+package com.management.tab.application.auth;
 
-import com.management.tab.config.auth.security.service.dto.LoggedInUserDto;
+import com.management.tab.application.auth.dto.LoggedInUserDto;
 import com.management.tab.config.auth.security.repository.UserSocialRepository;
 import com.management.tab.domain.user.User;
 import com.management.tab.domain.user.vo.RegistrationId;
