+++ src/main/java/com/management/tab/application/auth/dto/LoggedInUserDto.java
@@ -1,4 +1,4 @@
-package com.management.tab.config.auth.security.service.dto;
+package com.management.tab.application.auth.dto;
 
 public record LoggedInUserDto(Long id, String nickname, boolean isSignUp) {
 }
